Split builder wrapper tests into their own module
Move the builder wrapper test block into `wrapper_tests.rs` so the production
wrapper module stays below the repository file-size limit while preserving the
existing test module path and snapshots.
